requirements.txt
lxml==5.3.0
numpy==1.24.4
pandas==2.0.3
pillow==10.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
six==1.17.0
typing-extensions==4.12.2
tzdata==2024.2
